requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
alembic
psycopg2-binary
python-dotenv
requests
google-genai
python-multipart
python-jose[cryptography]
passlib[bcrypt]
ruff
black
pytest
pytest-asyncio
httpx
structlog
pydantic-settings

# Async DB drivers
aiosqlite
asyncpg

# RAG & Caching
numpy
redis
slowapi
pgvector
scikit-learn

# Monitoring
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.39.1
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp-proto-grpc

# Testing
pytest-asyncio
locust

reportlab>=4.0.0
# Notifications (P2-3)
twilio>=9.0.0
sendgrid>=6.12.0

# Testing
pytest-mock>=3.15.0
